chore(config): Decompose config handler architecture
